# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FRACAS & RAMS System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports/mdbf:
    get:
      operationId: getMDBFReport
      tags: [reports]
      summary: Get MDBF (Mean Distance Between Failures) report
      parameters:
        - in: query
          name: startDate
          schema:
            type: string
        - in: query
          name: endDate
          schema:
            type: string
      responses:
        "200":
          description: MDBF report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MDBFReport"
  /reports/mdbcf:
    get:
      operationId: getMDBCFReport
      tags: [reports]
      summary: Get MDBCF (Mean Distance Between Component Failures) report
      parameters:
        - in: query
          name: startDate
          schema:
            type: string
        - in: query
          name: endDate
          schema:
            type: string
      responses:
        "200":
          description: MDBCF report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MDBCFReport"
  /reports/mttr:
    get:
      operationId: getMTTRReport
      tags: [reports]
      summary: Get MTTR (Mean Time To Repair) report
      parameters:
        - in: query
          name: startDate
          schema:
            type: string
        - in: query
          name: endDate
          schema:
            type: string
      responses:
        "200":
          description: MTTR report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MTTRReport"
  /reports/availability:
    get:
      operationId: getAvailabilityReport
      tags: [reports]
      summary: Get fleet availability report
      parameters:
        - in: query
          name: startDate
          schema:
            type: string
        - in: query
          name: endDate
          schema:
            type: string
      responses:
        "200":
          description: Availability report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilityReport"
  /reports/pattern-failures:
    get:
      operationId: getPatternFailures
      tags: [reports]
      summary: Get pattern failure analysis
      parameters:
        - in: query
          name: windowMonths
          schema:
            type: number
      responses:
        "200":
          description: Pattern failure report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatternFailureReport"
...
components:
  schemas:
...
    MDBFReport:
      type: object
      properties:
        totalFleetDistance:
          type: number
        totalServiceFailures:
          type: number
        mdbf:
          type: number
        target:
          type: number
        compliance:
          type: boolean
        byTrain:
          type: array
          items:
            type: object
            properties:
              trainNumber:
                type: string
              distance:
                type: number
              serviceFailures:
                type: number
              mdbf:
                type: number
            required: [trainNumber, distance, serviceFailures, mdbf]
        trend:
          type: array
          items:
            type: object
            properties:
              period:
                type: string
              mdbf:
                type: number
              failures:
                type: number
            required: [period, mdbf, failures]
      required: [totalFleetDistance, totalServiceFailures, mdbf, target, compliance, byTrain, trend]
    MDBCFReport:
      type: object
      properties:
        bySystem:
          type: array
          items:
            type: object
            properties:
              systemCode:
                type: string
              systemName:
                type: string
              population:
                type: number
              totalDistance:
                type: number
              componentFailures:
                type: number
              mdbcf:
                type: number
            required: [systemCode, systemName, population, totalDistance, componentFailures, mdbcf]
      required: [bySystem]
    MTTRReport:
      type: object
      properties:
        overallMTTR:
          type: number
        target:
          type: number
        compliance:
          type: boolean
        bySystem:
          type: array
          items:
            type: object
            properties:
              systemCode:
                type: string
              systemName:
                type: string
              totalRepairs:
                type: number
              totalRepairTime:
                type: number
              mttr:
                type: number
            required: [systemCode, systemName, totalRepairs, totalRepairTime, mttr]
        distribution:
          type: array
          items:
            type: object
            properties:
              range:
                type: string
              count:
                type: number
            required: [range, count]
      required: [overallMTTR, target, compliance, bySystem, distribution]
    AvailabilityReport:
      type: object
      properties:
        overallAvailability:
          type: number
        target:
          type: number
        compliance:
          type: boolean
        scheduledHours:
          type: number
        unavailableHours:
          type: number
        byTrain:
          type: array
          items:
            type: object
            properties:
              trainNumber:
                type: string
              availability:
                type: number
              downtime:
                type: number
            required: [trainNumber, availability, downtime]
        trend:
          type: array
          items:
            type: object
            properties:
              period:
                type: string
              availability:
                type: number
            required: [period, availability]
      required: [overallAvailability, target, compliance, scheduledHours, unavailableHours, byTrain, trend]
    PatternFailureReport:
      type: object
      properties:
        windowMonths:
          type: number
        analysisDate:
          type: string
        patterns:
          type: array
          items:
            $ref: "#/components/schemas/PatternFailure"
      required: [windowMonths, analysisDate, patterns]
    PatternFailure:
      type: object
      properties:
        partNumber:
          type: string
        partDescription:
          type: string
        systemCode:
          type: string
        systemName:
          type: string
        occurrences:
          type: number
        populationAffected:
          type: number
        totalPopulation:
          type: number
        percentageAffected:
          type: number
        predictedRate:
          type: number
        actualRate:
          type: number
        rateExceedance:
          type: number
        isPattern:
          type: boolean
        patternType:
          type: string
        firstOccurrence:
          type: string
        lastOccurrence:
          type: string
      required:
        - partNumber
        - partDescription
        - systemCode
        - systemName
        - occurrences
        - populationAffected
        - totalPopulation
        - percentageAffected
        - isPattern
...